[RENAME/#181] 달력 모델 이름 수정
