+++ feature/src/main/java/com/terning/feature/calendar/month/model/DayModel.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Immutable
 import java.time.LocalDate
 
 @Immutable
-data class DayClass(
+data class DayModel(
     val date: LocalDate,
     val isOutDate: Boolean = false
 )
